squash: Rename file.
